Remove SignedMessage sending/handling from ProvingNode
