+++ src/states/common/mod.rs
@@ -10,11 +10,9 @@ mod base;
 mod bootstrapped;
 mod relocated;
 mod unapproved;
-mod unrelocated;
 
 pub use self::{
     base::Base, bootstrapped::Bootstrapped, relocated::Relocated, unapproved::Unapproved,
-    unrelocated::Unrelocated,
 };
 
 pub const USER_MSG_CACHE_EXPIRY_DURATION_SECS: u64 = 120;
