+++ src/states/common/unapproved.rs
@@ -8,13 +8,20 @@
 
 use super::Bootstrapped;
 use crate::{
+    chain::SectionInfo,
     error::RoutingError,
     id::PublicId,
-    messages::{HopMessage, RoutingMessage},
+    messages::{HopMessage, RoutingMessage, SignedMessage},
     routing_message_filter::FilteringResult,
+    routing_table::Authority,
+    time::Instant,
+    xor_name::XorName,
 };
+use std::collections::BTreeSet;
 
 pub trait Unapproved: Bootstrapped {
+    fn get_proxy_public_id(&self, proxy_name: &XorName) -> Result<&PublicId, RoutingError>;
+
     fn filter_hop_message(
         &mut self,
         hop_msg: HopMessage,
@@ -47,5 +54,40 @@ pub trait Unapproved: Bootstrapped {
         }
 
         Ok(Some(routing_msg))
+    }
+
+    fn send_routing_message_via_proxy(
+        &mut self,
+        routing_msg: RoutingMessage,
+        src_section: Option<SectionInfo>,
+        route: u8,
+        expires_at: Option<Instant>,
+    ) -> Result<(), RoutingError> {
+        if routing_msg.dst.is_client() && self.in_authority(&routing_msg.dst) {
+            return Ok(()); // Message is for us.
+        }
+
+        // Get PublicId of the proxy node
+        let proxy_pub_id = match routing_msg.src {
+            Authority::Client {
+                ref proxy_node_name,
+                ..
+            } => *self.get_proxy_public_id(proxy_node_name)?,
+            _ => {
+                error!("{} Source should be client if our state is Client", self);
+                return Err(RoutingError::InvalidSource);
+            }
+        };
+
+        let signed_msg = SignedMessage::new(routing_msg, self.full_id(), None)?;
+
+        if self.add_to_pending_acks(signed_msg.routing_message(), src_section, route, expires_at)
+            && !self.filter_outgoing_routing_msg(signed_msg.routing_message(), &proxy_pub_id, route)
+        {
+            let bytes = self.to_hop_bytes(signed_msg.clone(), route, BTreeSet::new())?;
+            self.send_or_drop(&proxy_pub_id, bytes, signed_msg.priority());
+        }
+
+        Ok(())
     }
 }
